ndg: add configuration discovery and initialization
